standardized sql folder and feed file names
